+++ calendar/CHANGELOG.md
@@ -8,6 +8,7 @@ new major version.
 
 ## Changes - `v1`
 
+- feat: respond with the expected object type for lambda invocations 2024-03-02
 - feat: prepare plans for casual computations on dates of a calendar 2024-02-29
 
 <!-- a collection of links -->